+++ jax/interpreters/pxla.py
@@ -30,6 +30,7 @@ from six.moves import reduce
 from .. import core
 from .. import ad_util
 from .. import tree_util
+from .. import lax_parallel as lp
 from .. import linear_util as lu
 from ..abstract_arrays import ShapedArray
 from ..util import partial, unzip2, concatenate, safe_map, prod
